KAN-107 [REFACTOR] Isolate Notice Recent Validation Logic with Business Rule Update
  * Isolate Notice Recent Validation logic to independent UseCase to enhance its reusability.
  * Apply updated business logic
    - 24-hours based validation to 48-hours based validation.
