+++ core/domain/src/main/java/com/doyoonkim/domain/usecases/FetchNoticesPerPage.kt
@@ -8,10 +8,6 @@ import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.catch
 import kotlinx.coroutines.flow.flowOn
 import kotlinx.coroutines.flow.transform
-import java.time.LocalDate
-import java.time.ZoneId
-import java.time.format.DateTimeFormatter
-import java.time.format.DateTimeParseException
 import javax.inject.Inject
 
 interface FetchNoticesPerPage {
@@ -20,6 +16,7 @@ interface FetchNoticesPerPage {
 
 class FetchNoticesPerPageImpl @Inject constructor(
     private val remoteRepository: NoticeRemoteRepository,
+    private val checkRecentNoticeUseCase: CheckRecentNotice,
     @IoDispatcher private val ioDispatcher: CoroutineDispatcher
 ) : FetchNoticesPerPage {
 
@@ -30,7 +27,7 @@ class FetchNoticesPerPageImpl @Inject constructor(
         }.transform { result ->
             result?.let {
                 if (lastNttId == 0) {
-                    emit(Result.success(it.checkRecentNotices()))
+                    emit(Result.success(checkRecentNoticeUseCase(it)))
                 } else {
                     emit(Result.success(it))
                 }
@@ -39,28 +36,4 @@ class FetchNoticesPerPageImpl @Inject constructor(
             /* Internal Error. */
             emit(Result.failure(it))
         }.flowOn(ioDispatcher)
-
-    private fun List<NoticeVO>.checkRecentNotices(): List<NoticeVO> {
-        // Retrieve Current Time
-        val current = System.currentTimeMillis()
-        // DateTimeFormatter
-        val format = DateTimeFormatter.ofPattern("yyyy-MM-dd")
-        // 24-hours in Millis
-        val dayInMillis = 86_400_000L
-        return this.map {
-            try {
-                val timeInMills = LocalDate.parse(it.timestamp, format)
-                    .atStartOfDay(ZoneId.of("Asia/Seoul"))
-                    .toInstant()
-                    .toEpochMilli()
-
-                it.copy(
-                    isRecent =  current - timeInMills <= dayInMillis
-                )
-            } catch (e: DateTimeParseException) {
-                // Unable to parse Date information. (Server-side formatting issue.)
-                it
-            }
-        }
-    }
 }
